refactor: Consolidate configuration tests and improve structure
Reorganizes configuration tests for better maintainability:
- Separates general config operations from first-run specific tests
- Introduces proper test setup and cleanup utilities
- Improves test structure with clear separation of concerns

Improves test coverage of configuration management functionality.
